flask-app/requirements.txt
face_recognition==1.3.0
Pillow==8.2.0
numpy==1.21.0
Flask==2.0.1
Werkzeug==2.0.1